requirements.txt
future~=1.0.0
torch~=2.3.1
numpy~=1.26.4
matplotlib~=3.8.4
pydub~=0.25.1
torchaudio~=2.3.1
scikit-learn~=1.4.2
speechbrain~=1.0.0
openai-whisper